aggro no longer triggers on scan

+++ act.informative.cpp
@@ -1466,7 +1466,7 @@ void look_at_room ( Character *ch, int ignore_brief )
 	if ( ch->desc && ch->desc->mxp )
 		update_mxp_map ( ch );
 
-	if ( KILL_ALL_ENABLED && PRF_FLAGGED ( ch, PRF_AGGRO ) )
+	if ( KILL_ALL_ENABLED && PRF_FLAGGED ( ch, PRF_AGGRO ) && view_room == IN_ROOM ( ch ) )
 	{
 		command_interpreter ( ch, ( char * ) "kill all" );
 	}
